==> notebooks/P1- Preprocessing Labels.ipynb <==
# OBJECTIVE:Deal with activities which overlap w.r.t timestamps in report

import pandas as pd
import datetime as dt
import numpy as np
import time

smart_glasses=pd.read_csv("glasses.csv")
smart_watch=pd.read_csv("smartwatch.csv")
smart_phone=pd.read_csv("smartphone.csv")
report=pd.read_csv("report.csv")

smart_glasses=smart_glasses.rename(columns={'DATE':'TIMESTAMP'})
smart_watch=smart_watch.rename(columns={'timestamp':'TIMESTAMP'})
smart_phone=smart_phone.rename(columns={'timestamp':'TIMESTAMP'})
report=report.rename(columns={'from':'FROM','to':'TO'})

smart_phone.TIMESTAMP =pd.to_datetime(smart_phone.TIMESTAMP)
smart_glasses.TIMESTAMP=pd.to_datetime(smart_glasses.TIMESTAMP)
smart_watch.TIMESTAMP=pd.to_datetime(smart_watch.TIMESTAMP)
report.FROM=pd.to_datetime(report.FROM)
report.TO=pd.to_datetime(report.TO)

l1=[]
for i in range(0,len(report)):
    l1.append([report.loc[i,'FROM'],report.loc[i,'TO']])
# l1=pdf1['from'].tolist()
s1=pd.Series(l1)
report['FROM_TO']=s1
report.head(3)

print("\n       SMARTGLASS\n\n")
print(smart_glasses.head(2))
print("\n       SMARTWATCH\n\n")
print(smart_watch.head(2))
print("\n       SMARTPHONE\n\n")
print(smart_watch.head(2))
print("\n       REPORT\n\n")
print(report.head(2))

print("\n       SMARTGLASS\n\n")
print(smart_glasses.dtypes)
print("\n       SMARTWATCH\n\n")
print(smart_watch.dtypes)
print("\n       SMARTPHONE\n\n")
print(smart_phone.dtypes)
print("\n       REPORT\n\n")
print(report.dtypes)

report['activity_type'].value_counts()

# CREATING TIMESTAMPS FOR FINDING OUT ACTIVITIES CORRESPONDING TO THEM


l1=[]

def time_append(arg):
    start_time=arg[0]
    end_time=arg[1]
    temp_time=pd.date_range(start=start_time,end=end_time,freq='1min')
    #THE TIME GAP BETWEEN EACH ROWS COULD BE ADJUSTED BY CHANGING THE VALUE OF 'freq' IN ABOVE FUNCTION
    for i in temp_time:
        l1.append(i)
    return(temp_time)
t=(report['FROM_TO'].map(time_append))
df={'FROM':pd.Series(l1)}

CHECK=pd.DataFrame(df)
CHECK['TO']=CHECK['FROM']+dt.timedelta(minutes=1)
CHECK['TIMESTAMP']=CHECK['FROM']+dt.timedelta(minutes=0.5)
#REMOVING DUPLICATE ROWS
CHECK.drop_duplicates(keep='first')
CHECK['FROM_TO'] = list(zip(CHECK['FROM'], CHECK['TO']))
CHECK.drop(['FROM','TO'],axis=1,inplace=True)


#REMOVING DUPLICATE ROWS
CHECK.drop_duplicates(keep='first')


print(CHECK.shape)
print(CHECK.head(3))

# Finding out the activities across the respective timestamps

def checkf1(arg):
    report['check']=False
    report['check']=((report['FROM']<=arg)&(report['TO']>arg))
    a=report['activity_type'][report['check']]
    a.drop_duplicates(keep='first')
    a=a.tolist()
    if(len(a)>0):
        b=a[0]
        for i in range(1,len(a)):
            b=b+"/"+a[i]
        return(b)
    return("null")

    
CHECK['activity']=CHECK['TIMESTAMP'].map(checkf1)
report.drop('check',axis=1,inplace=True)
CHECK

CHECK.sort_values('TIMESTAMP',inplace=True)
CHECK.drop_duplicates()

# l=[]
# def append1(arg):
#     if(len(l)==0):
#         l.append(arg)
#         return(0)
#     else:
#         if(arg!=l[len(l)-1]):
#             l.append(arg)
#             return(1)
#         else:
#             return(2)
# CHECK['TEMP']=CHECK['activity'].map(append1)
# CHECK.drop(['TEMP','TIMESTAMP'],axis=1,inplace=True)
# print(len(l))
# l

CHECK['TEMP'] = list(zip(CHECK['FROM_TO'], CHECK['activity']))
CHECK.head()


# Making it into the report format

F=[]
T=[]
A=[]

def append2(arg):
    if(len(A)>0):
        if(arg[1]!=A[len(A)-1]):
            F.append(arg[0][0])
            A.append(arg[1])
            T.append(arg[0][1])
            return(1)
        else:
            T[len(T)-1]=arg[0][1]
            return(2)
    else:
        F.append(arg[0][0])
        A.append(arg[1])
        T.append(arg[0][1])
        return(0)
CHECK['TEMP1']=CHECK['TEMP'].map(append2)
CHECK.drop(['TEMP1'],axis=1,inplace=True)
report.drop('FROM_TO',axis=1,inplace=True)

print(len(A),len(F),len(T))


def splitf(arg):
    a=arg.split(" ")[2]
    return(a[3:8])

df={'FROM':pd.Series(F),'TO':pd.Series(T),'activity':pd.Series(A)}
final_report=pd.DataFrame(df)
final_report=final_report[final_report['activity']!='null']
final_report['DURATION']=final_report['TO']-final_report['FROM']

final_report['DURATION']=final_report['DURATION'].astype(str)
final_report['DURATION']=final_report['DURATION'].map(splitf)


# ..split(" ")[2])
final_report.to_csv("final_report.csv",sep=",")
final_report


final_report.activity.value_counts()

len(final_report)
