==> notebooks/P3-Combining datasets.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

smart_glasses=pd.read_csv("data/glasses.csv")
smart_watch=pd.read_csv("data/F_smart_watch.csv")
smart_phone=pd.read_csv("data/F_smart_phone.csv")
report_df = pd.read_csv("data/final_report.csv")

smart_phone.head()

# Renaming the columns

smart_phone_columns_dict = {col:'Phone_'+col for col in smart_phone.columns if col!='timestamp'}
smart_phone_columns_dict

smart_phone.rename(columns=smart_phone_columns_dict, inplace=True)

smart_phone.head()

smart_watch.head()

smart_watch_columns_dict = {col:'Watch_'+col for col in smart_watch.columns if col!='timestamp'}
smart_watch_columns_dict

smart_watch.rename(columns=smart_watch_columns_dict, inplace=True)

smart_watch.columns

smart_glasses.columns

smart_glasses.head()

smart_glasses.drop(columns=['NUM'], inplace=True)

smart_glasses.rename(columns={"DATE":"timestamp"}, inplace=True)

smart_glasses_columns_dict = {col:'Glass_'+col for col in smart_glasses.columns if col!='timestamp'}
smart_glasses_columns_dict

smart_glasses.rename(columns=smart_glasses_columns_dict, inplace=True)
smart_glasses.columns

# Rounding the timestamp into seconds, Group wrt timestamp & extract mean value of the features

smart_watch.timestamp = pd.to_datetime(smart_watch.timestamp)
smart_phone.timestamp = pd.to_datetime(smart_phone.timestamp)
smart_glasses.timestamp = pd.to_datetime(smart_glasses.timestamp)

smart_watch['timestamp']=smart_watch['timestamp'].dt.round('1s')
smart_phone['timestamp']=smart_phone['timestamp'].dt.round('1s')
smart_glasses['timestamp']=smart_glasses['timestamp'].dt.round('1s')

print("smart_watch shape : ",smart_watch.shape)
print("smart_phone shape : ",smart_phone.shape)
print("smart_glasses shape : ",smart_glasses.shape)

smart_watch.groupby("timestamp").mean()

smart_watch = smart_watch.groupby("timestamp").mean()
smart_glasses = smart_glasses.groupby("timestamp").mean()
smart_phone = smart_phone.groupby("timestamp").mean()

smart_glasses.reset_index(inplace=True)
smart_watch.reset_index(inplace=True)
smart_phone.reset_index(inplace=True)

print("smart_watch shape : ",smart_watch.shape)
print("smart_phone shape : ",smart_phone.shape)
print("smart_glasses shape : ",smart_glasses.shape)

features_df = pd.merge(smart_watch, smart_glasses, on='timestamp', how='outer')
features_df = pd.merge(features_df, smart_phone, on='timestamp', how='outer')

features_df.shape

features_df.columns

features_df.to_csv("data/features_combined.csv", index=None)

# Merging the features with labels

report_df = pd.read_csv("final_report.csv")
report_df

report_df['FROM'] = pd.to_datetime(report_df['FROM'])
report_df['TO'] = pd.to_datetime(report_df['TO'])

# Create an empty DataFrame to store the expanded data
expanded_df = pd.DataFrame(columns=['timestamp', 'activity'])

# Iterate through each row in 'report_df' and expand the data for each second
for _, row in report_df.iterrows():
    from_time = row['FROM']
    to_time = row['TO']
    activity = row['activity']
    
    # Create a DatetimeIndex with 1-second frequency
    time_index = pd.date_range(from_time, to_time, freq='S')
    
    # Create a DataFrame with the expanded data for this row
    data = pd.DataFrame({'timestamp': time_index, 'activity': activity})
    
    # Append the data to the expanded DataFrame
    expanded_df = expanded_df.append(data)

# Reset the index of the expanded DataFrame
expanded_df.reset_index(drop=True, inplace=True)

# Print the resulting DataFrame
print(expanded_df)

expanded_df.activity.value_counts()

report_df = expanded_df
report_df.head()

final_df = features_df.merge(report_df, on="timestamp", how="inner")

final_df.shape

final_df.to_csv("data/combined_sensor_data_with_features.csv", index=None)
