==> notebooks/P4 Handling missing values- ffill, dropping and combiniing labels.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

data_df = pd.read_csv("data/combined_sensor_data_with_features.csv")
data_df.head()

# Convert 'timestamp' column to datetime type
data_df['timestamp'] = pd.to_datetime(data_df['timestamp'])

# Set 'timestamp' as the index for easy manipulation
data_df.set_index('timestamp', inplace=True)

data_df.columns

pd.DataFrame(data_df.isna().sum())

features = ['Watch_battery', 'Watch_pressure', 'Watch_step_counter',
       'Watch_step_detector', 'Watch_HEART_RATE_X', 'Watch_HEART_RATE_Y',
       'Watch_ACC_X', 'Watch_ACC_Y', 'Watch_ACC_Z', 'Watch_GRAVITY_X',
       'Watch_GRAVITY_Y', 'Watch_GRAVITY_Z', 'Watch_GYRO_X', 'Watch_GYRO_Y',
       'Watch_GYRO_Z', 'Watch_LA_X', 'Watch_LA_Y', 'Watch_LA_Z', 'Watch_MAG_X',
       'Watch_MAG_Y', 'Watch_MAG_Z', 'Watch_ORI_X', 'Watch_ORI_Y',
       'Watch_ORI_Z', 'Watch_ROTATION_VECTOR_A', 'Watch_ROTATION_VECTOR_B',
       'Watch_ROTATION_VECTOR_C', 'Watch_ROTATION_VECTOR_D',
       'Watch_ROTATION_VECTOR_E', 'Glass_ACC_X', 'Glass_ACC_Y', 'Glass_ACC_Z',
       'Glass_GYRO_X', 'Glass_GYRO_Y', 'Glass_GYRO_Z', 'Glass_EOG_L',
       'Glass_EOG_R', 'Glass_EOG_H', 'Glass_EOG_V', 'Phone_battery',
       'Phone_light', 'Phone_pressure', 'Phone_proximity',
       'Phone_step_counter', 'Phone_step_detector', 'Phone_ACC_X',
       'Phone_ACC_Y', 'Phone_ACC_Z', 'Phone_GRAVITY_X', 'Phone_GRAVITY_Y',
       'Phone_GRAVITY_Z', 'Phone_GYRO_X', 'Phone_GYRO_Y', 'Phone_GYRO_Z',
       'Phone_LA_X', 'Phone_LA_Y', 'Phone_LA_Z', 'Phone_MAG_X', 'Phone_MAG_Y',
       'Phone_MAG_Z', 'Phone_ORI_X', 'Phone_ORI_Y', 'Phone_ORI_Z',
       'Phone_AUDIO_W', 'Phone_AUDIO_X', 'Phone_AUDIO_Y', 'Phone_AUDIO_Z',
       'Phone_ROTATION_VECTOR_A', 'Phone_ROTATION_VECTOR_B',
       'Phone_ROTATION_VECTOR_C', 'Phone_ROTATION_VECTOR_D',
       'Phone_ROTATION_VECTOR_E']

for feature in features:
    # Forward fill missing values for the next 60 seconds
    data_df[feature] = data_df[feature].fillna(method='ffill', limit=60)
    
    data_df[feature] = data_df[feature].fillna(method='bfill', limit=60)

# Find missing values

missing_values = data_df.isna().mean()*100
missing_values

missing_values[missing_values>30]

columns_to_drop = missing_values[missing_values>30].index.tolist()
columns_to_drop

data_df.drop(columns=columns_to_drop, inplace=True)

data_df.columns

# Sensor data count wrt activities

activity_data_count = data_df.groupby('activity').count()
activity_data_count

# saving it locally for analysis
activity_data_count.to_csv("Activity_vs_SensorDataCount.csv")

len(data_df.activity.value_counts())

data_df.shape

labels_to_drop = ['At home/Eat','Football','Running','Watching TV/At home']

data_df["is_label_to_drop"] = data_df.activity.map(lambda x: x in labels_to_drop)
data_df = data_df[data_df.is_label_to_drop==False]

data_df.drop(columns=['is_label_to_drop'], inplace=True)

print(data_df.shape)

data_df.activity.value_counts()

relabel_dict = {"On bus stop":"On bus stop/Walk",
               "Took off glasses/Shop/Walk":"Shop/Walk",
               "In computer/Eat":"In computer",
               "Eat/In computer":"In computer"}

data_df.activity = data_df.activity.map(lambda x: relabel_dict.get(x,x))

data_df.activity.value_counts()

data_df.isna().sum()

data_df.columns

features = ['Phone_pressure', 'Phone_ACC_X', 'Phone_ACC_Y', 'Phone_ACC_Z',
       'Phone_GRAVITY_X', 'Phone_GRAVITY_Y', 'Phone_GRAVITY_Z', 'Phone_GYRO_X',
       'Phone_GYRO_Y', 'Phone_GYRO_Z', 'Phone_LA_X', 'Phone_LA_Y',
       'Phone_LA_Z', 'Phone_MAG_X', 'Phone_MAG_Y', 'Phone_MAG_Z',
       'Phone_ORI_X', 'Phone_ORI_Y', 'Phone_ORI_Z', 'Phone_AUDIO_W',
       'Phone_AUDIO_X', 'Phone_AUDIO_Y', 'Phone_AUDIO_Z',
       'Phone_ROTATION_VECTOR_A', 'Phone_ROTATION_VECTOR_B',
       'Phone_ROTATION_VECTOR_C', 'Phone_ROTATION_VECTOR_D',
       'Phone_ROTATION_VECTOR_E']

data_df.Phone_ACC_X.median()

for feature in features:
    median = data_df[feature].median()
    data_df[feature].fillna(median, inplace=True)

data_df.isna().sum()

data_df.to_csv("data/P4_engineered_data.csv")
